+++ include/quicksort_cpu_version.h
@@ -1,11 +1,11 @@
 #ifndef QUICKSORT_CPU_VERSION_H
 #define QUICKSORT_CPU_VERSION_H
 
-void quicksort(int* vec, const int start, const int end) ; 
+void quicksort(float* vec, const int start, const int end) ; 
 
-void swap(int* a, int* b) ; 
+void swap(float* a, float* b) ; 
 
-int partion(int*vec, const int start, const int end) ; 
+int partion(float*vec, const int start, const int end) ; 
 
 
 #endif
